ml_project/requirements.txt
antlr4-python3-runtime==4.9.3
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
DataSynthesizer==0.1.11
fonttools==4.38.0
hydra-core==1.2.0
idna==3.4
joblib==1.2.0
kaggle==1.5.12
kiwisolver==1.4.4
matplotlib==3.6.0
numpy==1.23.4
omegaconf==2.2.3
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pydantic==1.10.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.5
PyYAML==6.0
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
text-unidecode==1.3
threadpoolctl==3.1.0
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
mlflow==1.30.0
